requirements.txt
fastapi==0.115.0
uvicorn==0.30.6
faiss-cpu==1.8.0
pymongo==4.8.0
sentence-transformers==3.1.1
PyPDF2==3.0.1
langchain==0.3.0
pytest==8.3.2
requests==2.32.3
python-dotenv==1.0.1
python-multipart==0.0.9  # Added